requirements.txt
torch>=2.0.0
transformers>=4.40.0
sentence-transformers>=2.7.0
accelerate>=0.30.0
einops>=0.7.0
numpy>=1.24.0
tqdm>=4.66.0
safetensors>=0.4.0
huggingface-hub>=0.23.0
bitsandbytes>=0.43.0
scipy>=1.11.0
chromadb>=0.4.0
lightrag-hku>=0.1.0
networkx>=3.0
httpx>=0.25.0
openai>=1.0.0
google-generativeai>=0.8.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
